requirements.txt
python-dotenv
discord.py
requests
aiohttp
aiosqlite